backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
slowapi
email-validator
